提交代码

+++ Server/Src/GameServer/GameObject/BuffObject.cpp
@@ -108,7 +108,7 @@ BOOL CBuffObject::OnEffect()
 	return TRUE;
 }
 
-BOOL CBuffObject::OnUpdate( UINT32 dwTick )
+BOOL CBuffObject::OnUpdate( UINT64 dwTick )
 {
 	if(m_bOver)
 	{
