+++ Server/Src/GameServer/SceneLogic/SceneLogic_Base.cpp
@@ -101,7 +101,7 @@ BOOL SceneLogicBase::OnObjectDie(CSceneObject* pObject)
 	return TRUE;
 }
 
-BOOL SceneLogicBase::Update(UINT32 dwTick)
+BOOL SceneLogicBase::Update(UINT64 dwTick)
 {
 	if(CommonFunc::GetCurrTime() - m_pScene->GetStartTime() > m_dwLastTime)
 	{
